+++ src/server/services/__tests__/govuk-notify.test.ts
@@ -10,6 +10,7 @@ import {
   sendEditDetailsEmail,
   sendDataPublishedEmail,
   sendManualActionNotificationToPost,
+  sendEmails,
 } from "../../../server/services/govuk-notify";
 import { logger } from "../../../server/services/logger";
 import { NOTIFY } from "../../config";
@@ -522,5 +523,31 @@ describe("GOVUK Notify service:", () => {
       const settled = await sendManualActionNotificationToPost(1, "CHANGED_DETAILS");
       expect(settled).toEqual({ statusText: "Created" });
     });
+  });
+});
+
+describe("sendEmails", () => {
+  test("sendEmails calls sendEmail the correct amount of times", async () => {
+    const notifyClient = getNotifyClient();
+    const spy = jest.spyOn(notifyClient, "sendEmail");
+
+    await sendEmails("abc", ["test@gov.uk", "test@gov.uk"], {});
+
+    expect(spy).toBeCalledTimes(2);
+  });
+
+  test("sendEmails only rejects when all emails fail", async () => {
+    const notifyClient = getNotifyClient();
+    const spy = jest.spyOn(notifyClient, "sendEmail");
+    spy.mockRejectedValue("Error");
+    await expect(sendEmails("abc", ["test@gov.uk", "test@gov.uk"], {})).rejects.toThrow(AggregateError);
+  });
+
+  test("sendEmails returns resolved value when at least one email sends", async () => {
+    const notifyClient = getNotifyClient();
+    const spy = jest.spyOn(notifyClient, "sendEmail");
+    spy.mockResolvedValueOnce({ successText: "woo" }).mockRejectedValue("Error");
+
+    expect(await sendEmails("abc", ["test@gov.uk", "test@gov.uk"], {})).toEqual({ successText: "woo" });
   });
 });
